move elo code to new class
